build(core): Move custom ant tasks to build-definitions.xml
Partially from
https://github.com/LWJGL/lwjgl3/commit/8fd5f8dd95b23bdf9094f17e8428b6608544b775

This stops `BindingConfig` from being called repeatedly when chaining
tasks like we do in `ci_build_android.bash` which causes issues as it
sets `-Dbinding.core=true` after hardcoding its inclusion in the modules
list. This causes `core` to duplicate at the end of the property
`module.list` (because BindingConfig has no deduplciation logic for
that) which messes with `ModuleInfoGen` as it tries to move a file it
already moved before, that being the `module-info.class` file for
`core`.

We could have just repeatedly run the ant command per task but that's
boring and boilerplatey. This is better.